% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_fidelity.R
\name{energy_fidelity}
\alias{energy_fidelity}
\title{Calculate the energy function for fidelity to underlying data}
\usage{
energy_fidelity(y, f)
}
\arguments{
\item{y}{the underlying observed lattice grid}

\item{f}{the current configuration of the model for which to compare}
}
\value{
A measure of the discordance between the observed data and the model
}
\description{
The \code{energy_fidelity} function calculates the energy associated with
the difference between the underlying observed data and the current
configuration of the model.
}
\examples{
  y <- matrix(sample(c(-1, 1), size = 16, prob = c(0.6, 0.4)), nrow = 4)
  f <- matrix(sample(c(-1, 1), size = 16, prob = c(0.65, 0.35), nrow = 4))
  energy_fidelity()

}
\author{
Kevin Brosnan (\email{kevin.c.brosnan@gmail.com})
}
