% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_global.R
\name{energy_global}
\alias{energy_global}
\title{Global energy of lattice grid}
\usage{
energy_global(x, trend = -1)
}
\arguments{
\item{x}{the lattice grid for which the global energy is required}

\item{trend}{the global tendancy of the lattice grid}
}
\value{
the value of the global energy function
}
\description{
\code{energy_global} calculates the complete global energy for the lattice
grid. The tendency of the overall grid position is provided through
\code{trend}.
}
\examples{
  x <- matrix(sample(c(-1, 1), times = 16, prob = c(0.6, 0.4)), nrow = 4)
  energy_global(x)

}
\author{
Kevin Brosnan (\email(kevin.c.brosnan@ul.ie))
}
