% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gating}
\alias{gating}
\title{gating}
\usage{
gating(x, var.names = NULL)
}
\arguments{
\item{x}{data frame with variables to be gated}

\item{var.names}{the variables to be used for gating}
}
\description{
gating
}
