% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingfc.R
\docType{package}
\name{gatingfc-package}
\alias{gatingfc}
\alias{gatingfc-package}
\title{gatingfc: automated gating of cytometry data}
\description{
gatingfc provides a framework for the statsitical analysis, and particularly
the gating, of flow cytometry data.
}
\details{
It addresses four of the objectives of the flow cytometry analysis framework:

\itemize{
\item Normalisation
\item Outlier Removal
\item Gating
\item Cluster Labelling
}

To learn how to use gatingfc, please see the introductory vignettes:
`browseVignettes(package = "gatingfc")`
}
\author{
\strong{Maintainer}: Kevin Brosnan \email{kevin.c.brosnan@gmail.com}

}
