% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Generates a grid of count data}
\usage{
make_grid(x)
}
\arguments{
\item{x}{a two column matrix of flow cytometry varaibles or a vector
representing on cytometry variable.}
}
\value{
A square count matrix representing the observed data. Values greater
the 0 represent activity of cells, while 0 records an inactive cell location.
}
\description{
\code{make_grid} generates a lattice grid for the bivariate data inputted
by \code{x}.
}
\details{
This function produces a grid of bivariate count data observed in flow
 cytometry analysis. This data lies on a lattice support grid where cells
 show the bivariate counts.
}
\examples{
 x <- matrix(sample(c(0:7), size = 10, replace = TRUE), ncol = 2)
 y <- make_grid(x)

}
\author{
Kevin Brosnan \email{kevin.c.brosnan@gmail.com}
}
