% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood_moore.R
\name{neighbourhood_moore}
\alias{neighbourhood_moore}
\title{Moore Neighbourhood Algorithm}
\usage{
neighbourhood_moore(x, order)
}
\arguments{
\item{x}{the lattice grid for which a neighbourhood structure is required}

\item{order}{the order of the neighbourhood system}
}
\value{
a matrix, where each row corresponds to an observations neighbours
}
\description{
This functions computes the complete neighbourhood system of the qth order
for a lattice grid under the Moore algorithm
}
\examples{
 y <- matrix(sample(c(-1, 1), size = 16, prob = c(0.6, 0.4), replace = TRUE), nrow = 4)
 neigh <- neighbourhood_moore(x = y, order = 1)

}
\author{
Kevin Brosnan (\email{kevin.c.brosnan@gmail.com})
}
