% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{neighbourhoods}
\alias{neighbourhoods}
\title{Neighbourhood of complete lattice grid}
\usage{
neighbourhoods(x, order = 1, method = "Moore")
}
\arguments{
\item{x}{the lattice grid for which a neighbourhood structure is required}

\item{order}{the order of the neighbourhood system}

\item{method}{the neighbourhood method, 'Moore' or 'von Neumann'}
}
\value{
a matrix, where each row corresponds to an observations neighbours
}
\description{
This function provides a single wrapper to \code{neighbourhood_moore} and
\code{neighbourhood_neumann}. It is advised that this function be called to
find the neighbourhoods in this package, as it provides the most error checks
}
\examples{
 y <- matrix(sample(c(-1, 1), size = 16, prob = c(0.6, 0.4), replace = TRUE), nrow = 4)
 neigh <- neighbourhoods(x = y, order = 1, method = "Moore")

}
\author{
Kevin Brosnan (\email{kevin.c.brosnan@gmail.com})
}
