% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gating}
\alias{gating}
\title{gating}
\usage{
gating(x, var.names = NULL)
}
\arguments{
\item{x}{data frame with variables to be gated}

\item{var.names}{the variables to be used for gating}
}
\value{
a data frame with the model output specified as:
}
\description{
The \code{gating} function is the main component of the \code{gatingfc}
package. It combines many of the other functions within this package
and provides a wrapper function to automated the procedure of gating.
}
\author{
Kevin Brosnan (\email{kevin.c.brosnan@gmail.com})
}
